1.2.0 push
Fix the bug Bid price input text will be erased when bid panel is refreshed. The refresh happens when bid status changes, like price, winner, etc.
Fix the bug delete item causing delete multiple ones in one click
Fix the bug gold sharing is not calculated correctly.
Set the default raid size to 10 instead of 1 to prevent items got deal due to RL forget setting the raid size at the beginning.
Pause Bid button will do the work "End Bid" button did.
"End Bid" button now will close the bid. Any items still in bidding status will either goes to "Bought-In" (If no one bid) or "Deal" (to the current winner @ current price).

New Features:

Add a monitor window to master only. Master and either use /reunionlootmonitor to show it or click any item link when there are items in the bid panel.
Item link in master window now is a button.
Usage of monitor window:
Refresh will capture members in the raid.
Ping All will ping all clients
If the client window is not opened - names are in red
If the client window is open, but the version is old (need update the addon) - names are in yellow
If the client window is open and up to updated -- names are in green
When click any item link in the master window, the monitor window title changed to the item link
Name in Red -- means the players hasn't passed on this item
Name in Green -- means the player has passed on this item